Properly namespace V1 serializers
