--- app/serializers/ams/aid_station_serializer.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-class AidStationSerializer < BaseSerializer
-  attributes :id, :event_id, :split_id
-  link(:self) { api_v1_aid_station_path(object) }
-
-  belongs_to :event
-  belongs_to :split
-end
