--- app/serializers/ams/effort_with_times_row_serializer.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-class EffortWithTimesRowSerializer < BaseSerializer
-  attributes :event_short_name, :event_split_header_data
-  type :effort_with_times_row
-
-  has_one :effort_times_row
-end
