--- app/serializers/ams/event_spread_serializer.rb
@@ -1,16 +1,0 @@
-# frozen_string_literal: true
-
-class EventSpreadSerializer < BaseSerializer
-  attributes :name, :course_name, :organization_name, :event_start_time, :event_start_time_local, :display_style, :split_header_data
-  link(:self) { spread_api_v1_event_path(object.event) }
-
-  has_many :effort_times_rows
-
-  def event_start_time
-    object.start_time_local
-  end
-
-  def event_start_time_local
-    object.start_time_local
-  end
-end
