--- app/serializers/ams/raw_time_serializer.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-class RawTimeSerializer < BaseSerializer
-  attributes :id, :source, :absolute_time, :entered_time, :bib_number, :lap, :split_name, :sub_split_kind,
-             :data_status, :with_pacer, :remarks
-end
