+++ app/serializers/api/v1/course_serializer.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+module Api
+  module V1
+    class CourseSerializer < ::Api::V1::BaseSerializer
+      attributes :id, :name, :description, :editable, :locations, :track_points
+      link(:self) { api_v1_course_path(object) }
+
+      has_many :splits
+      belongs_to :organization
+
+      def locations
+        object.ordered_splits.select(&:has_location?).map do |split|
+          {id: split.id, base_name: split.base_name, latitude: split.latitude, longitude: split.longitude}
+        end
+      end
+    end
+  end
+end
