+++ app/serializers/api/v1/split_time_serializer.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module Api
+  module V1
+    class SplitTimeSerializer < ::Api::V1::BaseSerializer
+      attributes :id, :effort_id, :lap, :split_id, :bitkey, :absolute_time, :data_status, :pacer, :remarks
+      link(:self) { api_v1_split_time_path(object) }
+    end
+  end
+end
